+++ src/utils/validators.ts
@@ -18,7 +18,6 @@ import {
   ActionWithdraw,
   Nullable,
 } from './types';
-import {useQueries} from '@tanstack/react-query';
 
 export type TokenType =
   | 'ERC-20'
@@ -344,45 +343,4 @@ export async function updateVerification(
   } else {
     return result;
   }
-}
-
-/**
- *  This method is a Mock validation function until the real SDK functions are ready
- * @param address dao address
- * @returns an arrea of queries the indicates the status of verifications
- */
-export function updateValidation(address: string) {
-  // FIXME: remove this function and use the real SDK function
-  function getRandomInt(max: number) {
-    return Math.floor(Math.random() * max);
-  }
-
-  const verificationQueries = [
-    {
-      queryKey: ['pluginRegistry', address],
-      queryFn: () =>
-        new Promise(resolve => {
-          setTimeout(() => resolve(Boolean(getRandomInt(2))));
-        }),
-      enabled: Boolean(address),
-      retry: false,
-    },
-    {
-      queryKey: ['pluginSetupProcessor', address],
-      queryFn: () =>
-        new Promise(resolve => {
-          setTimeout(() => resolve(Boolean(getRandomInt(2))));
-        }),
-      enabled: Boolean(address),
-      retry: false,
-    },
-  ];
-
-  return useQueries({
-    queries: verificationQueries.map(config => {
-      return {
-        ...config,
-      };
-    }),
-  });
 }
